requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0
numpy<2
scipy>=1.11.0

# Medical imaging
nibabel>=5.1.0
scikit-learn>=1.3.0

# Augmentation
albumentations>=1.3.1

# SAM
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git

# Visualisation & demo
matplotlib>=3.7.0
Pillow>=10.0.0
gradio>=4.0.0

# Utilities
tqdm>=4.66.0
